Move LightSwitch class definition to switch_as_x/light.py
